requirements.txt
crewai
browser-use
langchain-openai
flask
python-dotenv
playwright
